tag urls missing backslash
If a blog post has no tags in the head/meta section (even if it doesn't have the word Tags: ,in fact) it shows up in all tag-link result pages. In other words, if I have pages with Tag: pico, and click on that tag-link, any pages with no tags whatsoever will show up in the resulting page, as well as, of course, the pages that have the pico tag.